# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "這是一個簡單的智能合約 API 服務",
        "title": "Simple Storage API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/storage/value": {
            "get": {
                "description": "從智能合約中獲取當前存儲的值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "獲取存儲的值",
                "responses": {
                    "200": {
                        "description": "成功返回存儲的值",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "在智能合約中設置新的值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "設置新的值",
                "parameters": [
                    {
                        "description": "要設置的新值",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SetValueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功設置值",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "請求格式錯誤",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.SetValueRequest": {
            "type": "object",
            "required": [
                "value"
            ],
            "properties": {
                "value": {
                    "type": "string",
                    "example": "42"
                }
            }
        }
    }
}